+++ ToDo.ApiService/Data/TodoList.cs
@@ -1,0 +1,10 @@
+namespace ToDo.ApiService.Data;
+
+public class TodoList
+{
+    public int Id { get; set; }
+    public required string Name { get; set; }
+    public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+
+    public List<TodoItem> Items { get; set; } = [];
+}