+++ ToDo.ApiService/Program.cs
@@ -1,4 +1,5 @@
 using ToDo.ApiService.Data;
+using ToDo.ApiService.Features.Lists;
 using ToDo.ApiService.Features.Todos;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -26,12 +27,21 @@ if (app.Environment.IsDevelopment())
 using (var scope = app.Services.CreateScope())
 {
     var db = scope.ServiceProvider.GetRequiredService<TodoDbContext>();
+    db.Database.EnsureDeleted();
     db.Database.EnsureCreated();
 }
 
 app.MapGet("/", () => Results.Redirect("/swagger"));
 
 // Map feature endpoints (literal segments before parameterized routes)
+GetAllLists.MapEndpoint(app);
+CreateList.MapEndpoint(app);
+DeleteList.MapEndpoint(app);
+GetTodosByList.MapEndpoint(app);
+CreateTodoInList.MapEndpoint(app);
+ToggleTodoInList.MapEndpoint(app);
+DeleteTodoInList.MapEndpoint(app);
+
 GetAllTodos.MapEndpoint(app);
 CreateTodo.MapEndpoint(app);
 ToggleTodo.MapEndpoint(app);